feat!: Unify spec domain classes as records (#548)
Getter signatures are now changed.

Errors are not converted since they need to extend RuntimeException etc.
Similarly the Request/Response objects are not converted either, as we
are evaluating moving them out of the spec package. However, first we
need to double check if they are still used in the client facing API.

Fixes #507 🦕

Replaces: #509