feat!: Trim io.a2a.apec package (#556)
Introduce a new jsonrpc-common module, and move the io.a2a.spec Request
and Response classes there since they are internal use now.

The jsonrpc-common module also gets the JsonUtil and the various
MappingExceptions.

The spec package exceptions were tied to JSON-RPC. We decided to keep
the `code` and `data` fields, which could be useful even when JSON-RPC
is not used. But we got rid of the JSONRPCError class, and repurposes
the A2AError interface to take its role.
Also renamed JSONRPCRequest and JSONRPCResponse to be
A2ARequest/-Response.

Make A2AError extend RuntimeException rather than Error.

---------

Signed-off-by: Emmanuel Hugonnet <ehugonne@redhat.com>
Co-authored-by: Emmanuel Hugonnet <ehugonne@redhat.com>